chore: refactor benchmark projects and add benchmark workflow
