+++ sandbox/Benchmark/Properties/launchSettings.json
@@ -1,0 +1,51 @@
+﻿{
+  "profiles": {
+    "Default": {
+      "commandName": "Project",
+      "commandLineArgs": "--filter Benchmark.IterateBenchmark.*",
+
+      // Available benchmarks:
+      // "commandLineArgs": "--filter Benchmark.IterateBenchmark.*",
+      // "commandLineArgs": "--filter Benchmark.LinqPerfBenchmarks.AggregateBy00.*",
+      // "commandLineArgs": "--filter Benchmark.LinqPerfBenchmarks.Count00.*",
+      // "commandLineArgs": "--filter Benchmark.LinqPerfBenchmarks.CountBy00.*",
+      // "commandLineArgs": "--filter Benchmark.LinqPerfBenchmarks.GroupBy00.*",
+      // "commandLineArgs": "--filter Benchmark.LinqPerfBenchmarks.Order00.*",
+      // "commandLineArgs": "--filter Benchmark.LinqPerfBenchmarks.Where00.*",
+      // "commandLineArgs": "--filter Benchmark.LinqPerfBenchmarks.Where01.*",
+      // "commandLineArgs": "--filter Benchmark.SimdRange.*",
+      // "commandLineArgs": "--filter Benchmark.SimdSelect.*",
+      // "commandLineArgs": "--filter Benchmark.LookupBattle.*",
+      // "commandLineArgs": "--filter Benchmark.LookupBattle.*",
+      // "commandLineArgs": "--filter Benchmark.ReadMeBenchmark.*",
+      // "commandLineArgs": "--filter Benchmark.ShuffleBench.*",
+      // "commandLineArgs": "--filter Benchmark.SimdAny.*",
+      // "commandLineArgs": "--filter Benchmark.SimdRange.*",
+      // "commandLineArgs": "--filter Benchmark.SimdSelect.*",
+      // "commandLineArgs": "--filter Benchmark.SimdSum.*",
+      // "commandLineArgs": "--filter Benchmark.SimdSumUnsigned.*",
+
+      "workingDirectory": "."
+    },
+    "Select Benchmark": {
+      "commandName": "Project",
+      "commandLineArgs": ""
+    },
+    "Run All Benchmarks": {
+      "commandName": "Project",
+      "commandLineArgs": "--filter *",
+    },
+    "Run All Benchmarks with --join": {
+      "commandName": "Project",
+      "commandLineArgs": "--filter * --join",
+    },
+    "List Benchmark": {
+      "commandName": "Project",
+      "commandLineArgs": "--list tree",
+    },
+    "Help": {
+      "commandName": "Project",
+      "commandLineArgs": "--help",
+    }
+  }
+}
